tests/wpt/tests/shadow-dom/focus/focus-scroll-under-delegatesFocus.html

<!DOCTYPE html>
<link rel="author" href="mailto:masonf@chromium.org">
<link rel="help" href="https://issues.chromium.org/issues/324112201">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<div id=container style="border:1px solid red; width: 100px;">
  <template shadowrootmode=open shadowrootdelegatesfocus>
    <slot></slot>
  </template>

  <a id=anchor href="#heading">anchor</a>
  <div style="height:2000px"></div>
  <h1 id=heading>Heading</h1>
</div>

<script>
promise_test(async (t) => {
  t.add_cleanup(() => container.remove());
  let scrolled = new Promise(resolve => {
    document.addEventListener('scrollend',resolve,{once:true});
  })
  await test_driver.click(anchor);
  await scrolled;

  scrolled = false;
  document.addEventListener('scroll',() => {
    scrolled = true;
  });
  await test_driver.click(heading);
  await new Promise(resolve => t.step_timeout(resolve, 500));
  assert_false(scrolled,'The document should not scroll');
},'delegatesFocus shouldn\'t cause extra focus steps');
</script>
